Update readme.txt
Decided to contribute a little bit back.  I've been using this plugin for years.  Problem was, people (mainly Store owners) kept asking me if it was "safe" to install since the version number hadn't been updated in a while.  So, I thought I'd add a bit to the ```==Description==``` and update the ```Tested up to: 4.9.1``` version number as well.